Revert "Add support for named mutex zones to combine cuboids together"
This reverts commit 7d0d3e08a995d6d316c3faa9b0c9306fbece0698.
